+++ flask_cod_facilito/app.py
@@ -2,7 +2,7 @@ from flask import Flask, render_template
 from config.config import DevelopmentConfig
 from flask_wtf import CSRFProtect
 from config.mail import mail
-from models.general import db
+from models import db
 from routes import (home_route,
                     user_route,
                     rol_route,
