+++ flask_cod_facilito/models/Rol.py
@@ -1,0 +1,11 @@
+from . import db
+from sqlalchemy import SmallInteger
+import datetime
+
+
+class Rol(db.Model):
+    __tablename__ = "rol"
+    id = db.Column(db.Integer, primary_key=True)
+    rol = db.Column(db.String(20), unique=True)
+    status = db.Column(SmallInteger, nullable=True, default=1)
+    create_date = db.Column(db.DateTime, default=datetime.datetime.now)
