+++ flask_cod_facilito/models/__init__.py
@@ -1,0 +1,8 @@
+from .general import db
+from .Comment import Comment
+from .Medias import Medias
+from .Message import Message
+from .Registers import Registers
+from .Rol import Rol
+from .TypesId import Types_id
+from .Users import User
