+++ flask_cod_facilito/paypalPayment.py
@@ -1,0 +1,2 @@
+PAYPAL_LOAD = 'wibmatup.pay.paypal/me/${username}'
+PAYPAL_CHARGE = '${}, ${}, ${}, ${}'
