+++ flask_cod_facilito/routes/forms.py
@@ -2,7 +2,7 @@ from flask import render_template, request, flash
 from flask import Blueprint, render_template, request, current_app, url_for, redirect, flash
 from utils.decorators.decorators import already_logged_in
 from forms.web_form import CreateForm
-from models.general import (db, User, Types_id, Medias, Registers)
+from models import (db, User, Types_id, Medias, Registers)
 from config.mail import mail, Message, MailConfig
 from smtplib import SMTPException, SMTPAuthenticationError
